Add joining and leaving events on venue page

+++ api/server.js
@@ -121,13 +121,28 @@ app.get('/venues/:id', (req, res, next) => {
   const { id } = req.params
   db.all(
     `
-      SELECT * FROM Location
+      SELECT 
+        Location.Longitude,
+        Location.Latitude,
+        Event.EventID,
+        Event.Name,
+        SportType.SportName,
+        Event.StartTime,
+        Event.EndTime,
+        Event.EventAddedTime,
+        IsAttending
+      FROM Location
       LEFT JOIN Event ON Location.LocationID = Event.LocationID
       LEFT JOIN SportType on Event.SportID = SportType.SportID
+      LEFT JOIN UserAttendingEvent on Event.EventID = UserAttendingEvent.EventID
+      LEFT JOIN (
+        SELECT 1 as IsAttending FROM UserAttendingEvent WHERE UserAttendingEvent.UserID = ?
+      ) ON Event.EventID = UserAttendingEvent.EventID
       WHERE Location.LocationID = ?
+      GROUP BY Event.EventID
       ORDER BY Event.EventAddedTime DESC
     `,
-    id,
+    [req.session.userID, id],
     (err, rows) => {
       if (err) console.error(err)
       // console.log('/venues/:id', rows)
@@ -147,7 +162,8 @@ app.get('/venues/:id', (req, res, next) => {
             SportName: row.SportName,
             StartTime: row.StartTime,
             EndTime: row.EndTime,
-            EventAddedTime: row.EventAddedTime
+            EventAddedTime: row.EventAddedTime,
+            IsAttending: row.IsAttending
           }
         })
       }
@@ -476,25 +492,43 @@ app.get('/SignedUpEvents/:id/Attending', (req, res, next) => {
 
 // Post request to deleteEvents for a specfic user
 app.post('/DeleteEvents', (req, res) => {
+  if (!req.session.userID) return res.json({
+    error: 'Not logged in'
+  })
+
   db.serialize(() => {
     if(req.body.TypeOfEvent == 'UserEvents'){
       db.run(`DELETE FROM Event WHERE EventID = ?`, req.body.EventID ,(err, row) => {
           if (err) {
             console.error(err.message);
+            res.json({
+              error: 'Could not delete event.'
+            })
           }
         });
     }
     if(req.body.TypeOfEvent == 'SignedUpEvents'){
-        db.run(`DELETE FROM UserAttendingEvent WHERE EventID = ?`, req.body.EventID ,(err, row) => {
-          if (err) {
-            console.error(err.message);
-          }
-        });
+        db.run(
+          `DELETE FROM UserAttendingEvent WHERE EventID = ? AND UserID = ?`,
+          [req.body.EventID, req.session.userID],
+          (err, row) => {
+            if (err) {
+              console.error(err)
+              res.json({
+                error: 'Could not leave event.'
+              })
+            }
+          });
+
         db.run(`UPDATE Event
                  SET PeopleAttending = PeopleAttending - 1
                  WHERE EventID = ?;`,req.body.EventID ,(err, row) => {
                 if (err) {
                   console.error(err.message);
+                } else {
+                  res.json({
+                    success: true
+                  })
                 }
         });
 
@@ -507,11 +541,20 @@ app.post('/DeleteEvents', (req, res) => {
 
 // Post request used to Join event
 app.post('/JoinEvent', (req, res) => {
+  if (!req.session.userID) {
+    return res.json({
+      error: 'Not logged in'
+    })
+  }
+
   db.serialize(() => {
     db.run(`INSERT INTO UserAttendingEvent (UserID, EventID)
-            VALUES (?,?);`,req.body.UserID, req.body.EventID ,(err, row) => {
+            VALUES (?,?);`,req.session.userID, req.body.EventID ,(err, row) => {
             if (err) {
               console.error(err.message);
+              res.json({
+                error: 'Could not join event'
+              })
             }
      });
      db.run(`UPDATE Event
@@ -519,7 +562,14 @@ app.post('/JoinEvent', (req, res) => {
               WHERE EventID = ?;`,req.body.EventID ,(err, row) => {
              if (err) {
                console.error(err.message);
+               return res.json({
+                 error: 'Could not join event'
+               })
              }
+
+             res.json({
+               success: true
+             })
       });
   });
   console.log("Joining Event ID  : ",req.body.EventID)
